requirements.txt
# Cors
django-cors-headers==2.4.0

# Core
pytz==2018.9
Django==2.1.5
django-configurations==2.1
gunicorn==19.9.0
newrelic==4.12.0.113

# For the persistence stores
psycopg2-binary==2.7.7
dj-database-url==0.5.0

# Model Tools
django-model-utils==3.1.2
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.9.1
Markdown==3.0.1
django-filter==2.1.0

# Django-Auth
django-allauth==0.37.1
django-rest-auth==0.9.3

# Developer Tools
ipdb==0.11
ipython==7.2.0
mkdocs==1.0.4
flake8==3.6.0

# Testing
mock==2.0.0
factory-boy==2.11.1
django-nose==1.4.6
nose-progressive==1.5.2
coverage==4.5.2
pytest-django

# Static and Media Storage
django-storages==1.7.1
boto3==1.9.83

#Django-phone-number
django-phonenumber-field
phonenumbers